My Codecademy Projects
This is my first time using GitHub, so hopefully this all uploads corectly. I just finished the Codecademy Learn C# course and have decided to publish the projects I did during that course as example code for now until I get my Unity Projects underway. There is an overall README and an individual README for each project. Feel free to read them if you have the time, but the code is not incredibly complex.

Tony